Add doActions.js
